functions/ai_monitoring/requirements.txt
firebase-admin==6.2.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
schedule==1.2.1
ultralytics==8.0.196 